Replace build tool to vite
