+++ .gitignore
@@ -8,9 +8,9 @@
 # testing
 /coverage
 
-# next.js
-/.next/
-/out/
+# vite
+/.vite/
+/dist/
 
 # production
 /build
@@ -30,8 +30,7 @@ yarn-error.log*
 .env.test.local
 .env.production.local
 
-# vercel
-.vercel
+# idea
+.idea
 
-#storybook
 storybook-static
