+++ lint-staged.config.js
@@ -1,5 +1,4 @@
 module.exports = {
-  "**/*.{js,jsx,ts,tsx,css}": "stylelint --fix",
   "**/*.{js,jsx,ts,tsx}": "eslint --fix",
   "**/*.{js,jsx,ts,tsx,css,md}": "prettier --write ",
   "**/*.{ts,tsx}": () => "tsc --noEmit --project tsconfig.json",
