+++ package.json
@@ -1,122 +1,95 @@
 {
-  "name": "@dataware-tools/app-launcher-next",
-  "version": "0.1.0",
+  "name": "app-launcher",
+  "version": "0.0.0",
+  "basePath": "/launcher/",
   "engines": {
     "node": "16.11.1",
     "npm": "8.x"
   },
-  "private": true,
+  "volta": {
+    "node": "16.11.1"
+  },
   "repository": "https://github.com/dataware-tools/app-launcher-next.git",
-  "homepage": "/launcher",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
+    "dev": "vite --open",
+    "build": "vite build; cp serve.json dist",
+    "start": "serve dist -p 3000 -s",
     "test": "run-s lint test:unit test:build test:visual-regression",
-    "lint": "run-s lint:eslint lint:stylelint lint:tsc",
+    "lint": "run-s lint:eslint lint:tsc",
     "lint:eslint": "eslint .",
-    "lint:stylelint": "stylelint ./src",
     "lint:tsc": "tsc --noEmit --project tsconfig.json",
     "test:unit": "jest",
-    "test:build": "cross-env NODE_OPTIONS=\"--max-old-space-size=4096\" run-s build",
     "test:visual-regression": "build-storybook --quiet && loki --requireReference --reactUri file:./storybook-static",
     "format": "prettier --write src/*",
+    "test:build": "cross-env NODE_OPTIONS=\"--max-old-space-size=4096\" run-s build",
     "storybook": "start-storybook -p 9009 -s public",
     "build-storybook": "build-storybook -s public",
     "prepare": "husky install"
   },
   "dependencies": {
     "@auth0/auth0-react": "^1.8.0",
-    "@dataware-tools/app-common": "^21.10.2",
+    "@dataware-tools/api-meta-store-client": "^0.5.1",
+    "@dataware-tools/api-permission-manager-client": "^0.2.1",
+    "@dataware-tools/app-common": "^21.11.1",
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.0.1",
-    "@mui/lab": "^5.0.0-alpha.49",
+    "@mui/icons-material": "^5.0.5",
+    "@mui/lab": "^5.0.0-alpha.51",
     "@mui/material": "^5.0.2",
     "@mui/styles": "^5.0.1",
-    "@types/js-yaml": "^4.0.3",
-    "cross-env": "^7.0.3",
     "immer": "^9.0.6",
-    "js-yaml": "^4.1.0",
-    "next": "^11.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.3.0",
-    "react-toastify": "^8.0.3",
-    "recoil": "^0.4.1",
+    "react-toastify": "^8.1.0",
+    "serve": "^12.0.1",
     "swr": "^1.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.14",
-    "@emotion/babel-plugin": "^11.2.0",
-    "@emotion/babel-preset-css-prop": "^11.2.0",
-    "@storybook/addon-actions": "^6.3.7",
-    "@storybook/addon-essentials": "^6.3.7",
-    "@storybook/addon-links": "^6.3.9",
-    "@storybook/react": "^6.3.7",
-    "@testing-library/jest-dom": "^5.12.0",
+    "@babel/core": "^7.16.0",
+    "@originjs/vite-plugin-commonjs": "^1.0.0",
+    "@rollup/plugin-commonjs": "^20.0.0",
+    "@storybook/addon-actions": "^6.3.12",
+    "@storybook/addon-essentials": "^6.3.12",
+    "@storybook/addon-links": "^6.3.12",
+    "@storybook/react": "^6.3.12",
+    "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.3.0",
-    "@types/jest": "^26.0.22",
-    "@types/node": "^14.14.37",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.8",
-    "@typescript-eslint/eslint-plugin": "^4.31.1",
-    "@typescript-eslint/parser": "^4.31.1",
-    "babel-loader": "^8.2.2",
+    "@types/jest": "^27.0.2",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
+    "@types/react-router-dom": "^5.3.0",
+    "@vitejs/plugin-react": "^1.0.7",
+    "babel-loader": "^8.2.3",
+    "cross-env": "^7.0.3",
     "eslint": "^7.32.0",
-    "eslint-config-next": "^11.1.2",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-config-standard": "^16.0.2",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-standard": "^16.0.3",
     "eslint-config-standard-jsx": "^10.0.0",
-    "eslint-config-standard-react": "^9.2.0",
-    "eslint-config-standard-with-typescript": "^20.0.0",
-    "eslint-import-resolver-typescript": "^2.4.0",
-    "eslint-plugin-compat": "^3.9.0",
+    "eslint-config-standard-react": "^11.0.1",
+    "eslint-config-standard-with-typescript": "^21.0.1",
+    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-plugin-compat": "^3.13.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-react": "^7.25.1",
+    "eslint-plugin-promise": "^5.1.1",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-unused-imports": "^1.1.4",
-    "husky": "^6.0.0",
-    "jest": "^27.2.4",
-    "lint-staged": "^10.5.4",
+    "eslint-plugin-standard": "^5.0.0",
+    "eslint-plugin-unused-imports": "^1.1.5",
+    "husky": "^7.0.4",
+    "jest": "^27.3.1",
+    "lint-staged": "^11.2.6",
     "loki": "^0.28.1",
-    "msw": "^0.28.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "2.2.1",
-    "semantic-ui-react": "^2.0.3",
-    "stylelint": "^13.13.1",
-    "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-recess-order": "^2.3.0",
-    "stylelint-config-recommended": "^5.0.0",
-    "stylelint-config-sass-guidelines": "^8.0.0",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-config-styled-components": "^0.1.1",
-    "stylelint-declaration-block-no-ignored-properties": "^2.3.0",
-    "stylelint-no-unsupported-browser-features": "^5.0.2",
-    "stylelint-processor-styled-components": "^1.10.0",
-    "ts-jest": "^27.0.5",
+    "prettier": "^2.4.1",
+    "ts-jest": "^27.0.7",
     "tsconfig-paths-webpack-plugin": "^3.5.1",
-    "typescript": "^4.4.3"
-  },
-  "browserslist": [
-    "last 2 Android versions",
-    "last 2 Chrome versions",
-    "last 2 ChromeAndroid versions",
-    "last 2 Edge versions",
-    "last 2 Electron versions",
-    "last 2 Firefox versions",
-    "last 2 FirefoxAndroid versions",
-    "maintained Node versions",
-    "last 2 Safari versions",
-    "last 2 Samsung versions"
-  ],
-  "volta": {
-    "node": "16.11.1"
+    "typescript": "^4.4.4",
+    "vite": "^2.5.4",
+    "vite-plugin-env-compatible": "^1.1.1",
+    "vite-tsconfig-paths": "^3.3.17"
   }
 }
