+++ serve.json
@@ -1,0 +1,8 @@
+{
+  "rewrites": [
+    {
+      "source": "index.html",
+      "destination": "launcher/index.html"
+    }
+  ]
+}
